+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## 0.26.0 (unreleased)
+## 0.26.0 (2018-04-04)
 
   * updated to Gradle 4.6
   * bundle license and dependency information